[lexical-compiler] Refactor: rename the package to @lexical/compiler with an entry per pass
## Description

`@lexical/pure-annotations` named one pass of what is becoming a build-time
compiler for Lexical code: alongside the tree-shaking pass in this PR there
is codegen for `importJSON`/`exportJSON` from a declarative schema. Both are
source-to-source transforms over the user's own definitions, delivered the
same way — a Vite/Rollup plugin plus plain transform functions — so they
belong in one package under a name that does not have to be renamed again
when the second one lands. It has never been published, so nothing outside
this branch refers to the old name.

Each pass is its own entry point, following the `@lexical/react` convention
(no `main`, one entry per `src/*.ts`, generated by `update-version`):

- `@lexical/compiler/PureAnnotations` — the tree-shaking pass
- `@lexical/compiler` — every pass, for a build that wants all of them

For a dev-only package the split does not save a consumer any shipped bytes,
and it cannot save install weight either (dependencies are package-level, so
a future pass's dependencies are installed either way — one heavy enough to
matter belongs in its own package). What it does is keep the boundary
between passes real from the first one, so codegen can be added without
rewriting the entry a consumer already imports.

Renamed with it: `packages/lexical-pure-annotations` →
`packages/lexical-compiler`, `LexicalPureAnnotations.mjs` →
`LexicalCompiler.mjs`, and the integration test to `compiler.test.mjs`. The
API keeps its feature names — `pureAnnotations()`, `transformPureAnnotations`
— which is what makes room for `codegen()` beside them. The root
`LexicalCompiler.js` www stub is gone: `create-www-stubs` writes one only for
single-module packages.

## Test plan

### Before

```
$ node -e "…getPackageBuildDefinition().modules"
build modules: [ 'LexicalCompiler <- index.ts' ]
```

One bundle, because a package with `main` is treated as single-entry.

### After

```
$ node -e "…getPackageBuildDefinition().modules"
build modules: [
  'LexicalCompiler <- index.ts',
  'LexicalCompilerPureAnnotations <- PureAnnotations.ts'
]

$ node scripts/build.mjs && node scripts/build.mjs --prod
DEV=0
PROD=0

$ ls packages/lexical-compiler/dist | grep PureAnnotations
LexicalCompilerPureAnnotations.dev.js
LexicalCompilerPureAnnotations.dev.mjs
LexicalCompilerPureAnnotations.js
LexicalCompilerPureAnnotations.mjs
LexicalCompilerPureAnnotations.node.mjs
PureAnnotations.d.ts

$ npx vitest --project unit --project scripts-unit --no-watch
 Test Files  266 passed (266)
      Tests  5266 passed | 1 skipped (5267)

$ npx vitest --project integration --no-watch
 Test Files  4 passed (4)
      Tests  190 passed (190)

$ pnpm run ci-check
No errors!
```

The integration suite covers the rename end to end: the `link:`ed
source-mode fixture installs the package by its new name and imports the
plugin from `@lexical/compiler/PureAnnotations`.
