+++ .flowconfig
@@ -28,6 +28,8 @@ module.name_mapper='^@lexical/code$' -> '<PROJECT_ROOT>/packages/lexical-code/fl
 module.name_mapper='^@lexical/code-core$' -> '<PROJECT_ROOT>/packages/lexical-code-core/flow/LexicalCodeCore.js.flow'
 module.name_mapper='^@lexical/code-prism$' -> '<PROJECT_ROOT>/packages/lexical-code-prism/flow/LexicalCodePrism.js.flow'
 module.name_mapper='^@lexical/code-shiki$' -> '<PROJECT_ROOT>/packages/lexical-code-shiki/flow/LexicalCodeShiki.js.flow'
+module.name_mapper='^@lexical/compiler$' -> '<PROJECT_ROOT>/packages/lexical-compiler/flow/LexicalCompiler.js.flow'
+module.name_mapper='^@lexical/compiler/PureAnnotations$' -> '<PROJECT_ROOT>/packages/lexical-compiler/flow/LexicalCompilerPureAnnotations.js.flow'
 module.name_mapper='^@lexical/devtools-core$' -> '<PROJECT_ROOT>/packages/lexical-devtools-core/flow/LexicalDevtoolsCore.js.flow'
 module.name_mapper='^@lexical/dragon$' -> '<PROJECT_ROOT>/packages/lexical-dragon/flow/LexicalDragon.js.flow'
 module.name_mapper='^@lexical/eslint-plugin$' -> '<PROJECT_ROOT>/packages/lexical-eslint-plugin/flow/LexicalEslintPlugin.js.flow'
@@ -54,7 +56,6 @@ module.name_mapper='^@lexical/mdast$' -> '<PROJECT_ROOT>/packages/lexical-mdast/
 module.name_mapper='^@lexical/offset$' -> '<PROJECT_ROOT>/packages/lexical-offset/flow/LexicalOffset.js.flow'
 module.name_mapper='^@lexical/overflow$' -> '<PROJECT_ROOT>/packages/lexical-overflow/flow/LexicalOverflow.js.flow'
 module.name_mapper='^@lexical/plain-text$' -> '<PROJECT_ROOT>/packages/lexical-plain-text/flow/LexicalPlainText.js.flow'
-module.name_mapper='^@lexical/pure-annotations$' -> '<PROJECT_ROOT>/packages/lexical-pure-annotations/flow/LexicalPureAnnotations.js.flow'
 module.name_mapper='^@lexical/react/ExtensionComponent$' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalExtensionComponent.js.flow'
 module.name_mapper='^@lexical/react/LexicalAutoEmbedPlugin$' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalAutoEmbedPlugin.js.flow'
 module.name_mapper='^@lexical/react/LexicalAutoFocusPlugin$' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalAutoFocusPlugin.js.flow'
