+++ eslint.config.mjs
@@ -452,13 +452,13 @@ export default [
 
   // Override: the /* @__PURE__ */ annotations on module-scope calls to the
   // side-effect-free lexical factories are injected at build time by
-  // @lexical/pure-annotations, so they do not belong in the sources. The
+  // @lexical/compiler, so they do not belong in the sources. The
   // rule is autofixable, which is how a branch written before the transform
   // existed migrates: `pnpm run lint:fix`. Annotations on anything else
   // (a third-party factory, a call inside a function body) are untouched.
   {
     files: ['packages/**', 'examples/**', 'dev-examples/**'],
-    ignores: ['packages/lexical-pure-annotations/**'],
+    ignores: ['packages/lexical-compiler/**'],
     rules: {
       '@lexical/internal/no-pure-annotation': ERROR,
     },
