+++ packages/lexical-compiler/flow/LexicalCompiler.js.flow
@@ -8,5 +8,5 @@
  */
 
 /**
- * LexicalPureAnnotations
+ * LexicalCompiler
  */
