+++ packages/lexical-compiler/flow/LexicalCompilerPureAnnotations.js.flow
@@ -4,8 +4,9 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
+ * @flow strict
  */
 
-'use strict';
-
-module.exports = require('./dist/LexicalPureAnnotations.js');
+/**
+ * LexicalCompilerPureAnnotations
+ */
