+++ packages/lexical-compiler/src/__tests__/unit/strict.test.ts
