+++ packages/lexical-compiler/src/__tests__/unit/treeShaking.test.ts
