+++ packages/lexical-devtools/tsconfig.json
@@ -16,6 +16,10 @@
       "@lexical/code-core": ["../lexical-code-core/src/index.ts"],
       "@lexical/code-prism": ["../lexical-code-prism/src/index.ts"],
       "@lexical/code-shiki": ["../lexical-code-shiki/src/index.ts"],
+      "@lexical/compiler": ["../lexical-compiler/src/index.ts"],
+      "@lexical/compiler/PureAnnotations": [
+        "../lexical-compiler/src/PureAnnotations.ts"
+      ],
       "@lexical/devtools-core": ["../lexical-devtools-core/src/index.ts"],
       "@lexical/dragon": ["../lexical-dragon/src/index.ts"],
       "@lexical/eslint-plugin": ["../lexical-eslint-plugin/src/index.ts"],
@@ -54,7 +58,6 @@
       "@lexical/offset": ["../lexical-offset/src/index.ts"],
       "@lexical/overflow": ["../lexical-overflow/src/index.ts"],
       "@lexical/plain-text": ["../lexical-plain-text/src/index.ts"],
-      "@lexical/pure-annotations": ["../lexical-pure-annotations/src/index.ts"],
       "@lexical/react/ExtensionComponent": [
         "../lexical-react/src/ExtensionComponent.tsx"
       ],
