--- packages/lexical-pure-annotations/package.json
@@ -1,81 +1,0 @@
-{
-  "name": "@lexical/pure-annotations",
-  "description": "Build-time /* @__PURE__ */ annotation injection for the side-effect-free Lexical factories",
-  "keywords": [
-    "lexical",
-    "editor",
-    "vite-plugin",
-    "rollup-plugin",
-    "tree-shaking"
-  ],
-  "version": "0.49.0",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/facebook/lexical.git",
-    "directory": "packages/lexical-pure-annotations"
-  },
-  "main": "./dist/LexicalPureAnnotations.js",
-  "bugs": {
-    "url": "https://github.com/facebook/lexical/issues"
-  },
-  "homepage": "https://lexical.dev/docs/packages/lexical-pure-annotations",
-  "dependencies": {
-    "@babel/parser": "^8.0.0",
-    "magic-string": "^0.30.21"
-  },
-  "devDependencies": {
-    "rollup": "^4.61.0"
-  },
-  "sideEffects": false,
-  "module": "./dist/LexicalPureAnnotations.mjs",
-  "exports": {
-    ".": {
-      "source": "./src/index.ts",
-      "import": {
-        "types@<5.2": "./dist/typescript-too-old.d.ts",
-        "types": "./dist/index.d.ts",
-        "development": "./dist/LexicalPureAnnotations.dev.mjs",
-        "production": "./dist/LexicalPureAnnotations.prod.mjs",
-        "node": "./dist/LexicalPureAnnotations.node.mjs",
-        "default": "./dist/LexicalPureAnnotations.mjs"
-      },
-      "require": {
-        "types@<5.2": "./dist/typescript-too-old.d.ts",
-        "types": "./dist/index.d.ts",
-        "development": "./dist/LexicalPureAnnotations.dev.js",
-        "production": "./dist/LexicalPureAnnotations.prod.js",
-        "default": "./dist/LexicalPureAnnotations.js"
-      }
-    }
-  },
-  "types": "./dist/typescript-too-old.d.ts",
-  "typesVersions": {
-    "*": {
-      "*": [
-        "./dist/typescript-too-old.d.ts"
-      ]
-    }
-  },
-  "peerDependencies": {
-    "typescript": ">=5.2"
-  },
-  "peerDependenciesMeta": {
-    "typescript": {
-      "optional": true
-    }
-  },
-  "files": [
-    "dist",
-    "src",
-    "!src/__tests__",
-    "!src/__bench__",
-    "!src/__mocks__",
-    "!src/**/*.test.ts",
-    "!src/**/*.test.tsx",
-    "!src/**/*.bench.ts",
-    "!src/**/*.bench.tsx",
-    "README.md",
-    "LICENSE"
-  ]
-}
