+++ scripts/__tests__/integration/fixtures/lexical-link-source-mode/package.json
@@ -14,7 +14,7 @@
     "lexical": "link:../../../../../packages/lexical"
   },
   "devDependencies": {
-    "@lexical/pure-annotations": "link:../../../../../packages/lexical-pure-annotations",
+    "@lexical/compiler": "link:../../../../../packages/lexical-compiler",
     "typescript": "^5.9.2",
     "vite": "^7.3.2"
   },
