+++ scripts/vite/lexicalMonorepoPlugin.ts
@@ -9,7 +9,7 @@ import {defineConfig, mergeConfig, type Plugin} from 'vite';
 
 // Imported from source (rather than by package name) so that a monorepo
 // vite config loads before the packages have been built.
-import {pureAnnotations} from '../../packages/lexical-pure-annotations/src/LexicalPureAnnotations.mjs';
+import {pureAnnotations} from '../../packages/lexical-compiler/src/LexicalCompiler.mjs';
 import viteModuleResolution from './viteModuleResolution';
 
 export default function lexicalMonorepoPlugin(): Plugin[] {
